% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{market}
\alias{market}
\title{Stock Market}
\format{
Numeric data frame with 103 rows and 5 columns.

\describe{
\item{JPMorgan_Bank}{}
\item{Citibank_Bank}{}
\item{WellsFargo_Bank}{}
\item{RoyalDutchShell_Oil}{}
\item{ExxonMobil_Oil}{}
}
}
\source{
\url{www.prenhall.com/statistics}
}
\usage{
market
}
\description{
The weekly rates of return for five stocks (p =5) listed on
the New York Stock Exchange are given.
}
\details{
Each observation is calculated as the price
variation over a week, relative to the price at the
start of the week. 103 consecutive rows are given.
}
\keyword{datasets}
