% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes_mdscal.R
\name{procrustes_mdscal}
\alias{procrustes_mdscal}
\title{procrustes_mdscal}
\usage{
procrustes_mdscal(
  X0,
  k = 2,
  diss,
  power = NULL,
  wght = NULL,
  l = 400,
  c = 4,
  n_cores = 1,
  seed = NULL
)
}
\arguments{
\item{X0}{A matrix or data-frame containing \eqn{n} observations of the
explanatory variables, with dimension \eqn{p} of at least 2.}

\item{k}{Desired dimension in the reduced space, default is 2.}

\item{diss}{Type of metric to produce the dissimilarities between observations.
\texttt{minkowski} works with Minkowski metric of order \texttt{power},
and \texttt{gower} works with Gower metric, using specific weights for variables.}

\item{power}{Power of the Minkowski metric. Default is 2 (Euclidean).}

\item{wght}{A vector of \eqn{p} weights, for the original variables in the
Gower metric. Default is 1 for each variable.}

\item{l}{Maximum size of the sub-groups in the splitting procedure. Default is 400.}

\item{c}{Number of alignment points, common to all the sub-groups. Default is 4.}

\item{n_cores}{Number of CPU cores to be used in the parallelization. Default is 1.
This option is not available in Windows.}

\item{seed}{A seed for the random selection of points to split the data.}
}
\value{
A data matrix of \eqn{n} rows and \eqn{k} columns, with the coordinates
of points in the reduced space.
}
\description{
Fast multidimensional scaling with divide and conquer
strategy using Procrustes transformations.
}
\details{
This functions implements DCO-MDS (\textit{Divide and conquer multidimensional scaling}))
with Procrustes transformations to align the different groups.
The scaling method is Kruskal distance scaling by minimizing Stress-1 with the SMACOF algorithm,
using splines as a mapping procedure of the dissimilarities.
The results are rotated to the principal axis. Two options
are available for the dissimilarities: Minkowski metric and Gower metric.
There is a parallelization of the main procedure of scaling and aligning the sub-groups.
}
\examples{
x <- MASS::mvrnorm(n=5000,mu=c(0,0,0,0),diag(c(1,2,3,4)))

Z <- procrustes_mdscal(x, k=2,diss='minkowski', power=1.5, seed = 16497)
plot(Z, pch=20, main = 'Fast multidimensional scaling k=2')

#' @references
Borg, I., & Groenen, P. J. F. (2005). Modern multidimensional scaling: theory and applications. Springer Science & Business Media.
Pedro Delicado y Cristian Pachon-Garcia. “Multidimensional Scaling for Big Data”. En: arXiv preprint arXiv:2007.11919 (2020).

}
