% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtools-package.R
\docType{package}
\name{mvtools-package}
\alias{mvtools}
\alias{mvtools-package}
\title{mvtools: Multivariate Analysis Tools}
\description{

}
\author{
\strong{Maintainer}: Pedro Cosatto \email{pcosatto@fi.uba.ar} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
